+++ modules/nunjucks-render.js
@@ -1,6 +1,6 @@
-traverse = require('./traverse')
+const traverse = require('./traverse')
 
-###*
+/**
  * Render input with Nunjucks
  * @param  {object} env     Nunjucks environment to render with
  * @param  {object} context Nunjucks context
@@ -9,9 +9,9 @@ traverse = require('./traverse')
  * @example
  *   render(env, { testVar: 'var value' }, '{{ testVar }}') -> 'var value'
  *   render(env, { testVar: 'var value' }, { ojb: { inner: '{{ testVar }}' } }) -> { ojb: { inner: 'var value' } }
-###
+*/
 module.exports = (env, context, input) => traverse(
   input,
   (tmpl) => env.renderString(tmpl, context),
-  (string) => string.includes('{{') or string.includes('{%')
+  (string) => string.includes('{{') || string.includes('{%')
 )